requirements.txt
pandas==2.2.2
tqdm==4.66.4
nltk==3.8.1
tiktoken==0.7.0
gpt4all==2.7.0
requests==2.32.3
openpyxl==3.1.5
evaluate==0.4.2
sentence-transformers==3.0.1
rouge_score==0.1.2
sentencepiece==0.2.0
protobuf==5.28.0
xlsxwriter==3.2.0
openai==1.45.1
vllm
python-dotenv
datasets
transformers
torch
accelerate